+++ kord/src/ml/base/gather.rs
@@ -17,7 +17,7 @@ use super::helpers::save_kord_item;
 
 /// Gather a sample from the microphone and save it to disk.
 #[coverage(off)]
-pub fn gather_sample(destination: impl AsRef<Path>, length_in_seconds: u8) -> Void {
+pub fn gather_sample(destination: impl AsRef<Path>, length_in_seconds: f32) -> Void {
     println!("Listening ...");
 
     let audio_data = futures::executor::block_on(get_audio_data_from_microphone(length_in_seconds))?;
